requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.3
django-appconf==1.0.4
django-audit-log==0.7.0
django-auditlog==0.4.7
django-compressor==2.4
django-countries==6.1.2
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-jsonfield==1.4.0
django-mathfilters==1.0.0
django-tenants==3.0.1
djangorestframework==3.10.3
flake8==3.6.0
idna==2.9
mccabe==0.6.1
ndg-httpsclient==0.5.1
Pillow==7.0.0
psycopg2==2.8.5
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.4.0
pycparser==2.20
pycryptodomex==3.9.7
pyflakes==2.0.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python-dateutil==2.6.0
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
stream-django==1.6.1
stream-python==3.1.1
stripe==2.45.0
urllib3==1.25.8
